# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CrossList Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ListingCondition = typeof ListingCondition[keyof typeof ListingCondition];


export const ListingCondition = {
  new_with_tags: 'new_with_tags',
  new_without_tags: 'new_without_tags',
  excellent: 'excellent',
  good: 'good',
  fair: 'fair',
} as const;

export type ListingStatus = typeof ListingStatus[keyof typeof ListingStatus];


export const ListingStatus = {
  draft: 'draft',
  published: 'published',
  sold: 'sold',
  archived: 'archived',
} as const;

export interface Listing {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  brand?: string | null;
  /** @nullable */
  size?: string | null;
  condition?: ListingCondition;
  status: ListingStatus;
  platforms?: string[];
  imageUrls?: string[];
  /** @nullable */
  poshmarkDescription?: string | null;
  /** @nullable */
  depopDescription?: string | null;
  /** @nullable */
  mercariDescription?: string | null;
  /** @nullable */
  soldPrice?: number | null;
  /** @nullable */
  soldAt?: string | null;
  createdAt: string;
  updatedAt?: string;
}

export type ListingInputCondition = typeof ListingInputCondition[keyof typeof ListingInputCondition];


export const ListingInputCondition = {
  new_with_tags: 'new_with_tags',
  new_without_tags: 'new_without_tags',
  excellent: 'excellent',
  good: 'good',
  fair: 'fair',
} as const;

export interface ListingInput {
  title: string;
  description?: string;
  price: number;
  originalPrice?: number;
  category?: string;
  brand?: string;
  size?: string;
  condition: ListingInputCondition;
  platforms?: string[];
  imageUrls?: string[];
}

export type ListingUpdateCondition = typeof ListingUpdateCondition[keyof typeof ListingUpdateCondition];


export const ListingUpdateCondition = {
  new_with_tags: 'new_with_tags',
  new_without_tags: 'new_without_tags',
  excellent: 'excellent',
  good: 'good',
  fair: 'fair',
} as const;

export type ListingUpdateStatus = typeof ListingUpdateStatus[keyof typeof ListingUpdateStatus];


export const ListingUpdateStatus = {
  draft: 'draft',
  published: 'published',
  sold: 'sold',
  archived: 'archived',
} as const;

export interface ListingUpdate {
  title?: string;
  description?: string;
  price?: number;
  originalPrice?: number;
  category?: string;
  brand?: string;
  size?: string;
  condition?: ListingUpdateCondition;
  status?: ListingUpdateStatus;
  platforms?: string[];
  imageUrls?: string[];
  poshmarkDescription?: string;
  depopDescription?: string;
  mercariDescription?: string;
}

export interface ListingStats {
  total: number;
  draft: number;
  published: number;
  sold: number;
  archived: number;
  totalRevenue: number;
  avgPrice: number;
  recentActivity: Listing[];
}

export type PublishRequestPlatformsItem = typeof PublishRequestPlatformsItem[keyof typeof PublishRequestPlatformsItem];


export const PublishRequestPlatformsItem = {
  poshmark: 'poshmark',
  depop: 'depop',
  mercari: 'mercari',
} as const;

export interface PublishRequest {
  platforms: PublishRequestPlatformsItem[];
}

export interface PlatformPublishResult {
  platform: string;
  success: boolean;
  /** @nullable */
  url?: string | null;
  /** @nullable */
  error?: string | null;
}

export interface PublishResult {
  results: PlatformPublishResult[];
}

export interface MarkSoldInput {
  soldPrice: number;
}

export interface RemoveBgInput {
  imageBase64: string;
  filename?: string;
}

export interface RemoveBgResult {
  imageBase64: string;
  filename: string;
}

export interface SoldItem {
  title: string;
  price: number;
  platform: string;
  soldDate: string;
  /** @nullable */
  condition?: string | null;
}

export interface SoldPriceData {
  query: string;
  avgPrice: number;
  minPrice: number;
  maxPrice: number;
  sampleCount: number;
  suggestedPrice?: number;
  items: SoldItem[];
}

export type GenerateDescriptionInputPlatformsItem = typeof GenerateDescriptionInputPlatformsItem[keyof typeof GenerateDescriptionInputPlatformsItem];


export const GenerateDescriptionInputPlatformsItem = {
  poshmark: 'poshmark',
  depop: 'depop',
  mercari: 'mercari',
} as const;

export interface GenerateDescriptionInput {
  title: string;
  category?: string;
  brand?: string;
  size?: string;
  condition?: string;
  notes?: string;
  platforms?: GenerateDescriptionInputPlatformsItem[];
}

export interface GeneratedDescriptions {
  /** @nullable */
  poshmark?: string | null;
  /** @nullable */
  depop?: string | null;
  /** @nullable */
  mercari?: string | null;
}

export interface Platform {
  name: string;
  displayName: string;
  connected: boolean;
  color: string;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  listingCount?: number | null;
}

export type ListListingsParams = {
status?: ListListingsStatus;
platform?: string;
};

export type ListListingsStatus = typeof ListListingsStatus[keyof typeof ListListingsStatus];


export const ListListingsStatus = {
  draft: 'draft',
  published: 'published',
  sold: 'sold',
  archived: 'archived',
} as const;

export type GetSoldPricesParams = {
query: string;
category?: string;
condition?: string;
};