# lib/api-zod/src/generated/types/generateDescriptionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CrossList Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { GenerateDescriptionInputPlatformsItem } from './generateDescriptionInputPlatformsItem';

export interface GenerateDescriptionInput {
  title: string;
  category?: string;
  brand?: string;
  size?: string;
  condition?: string;
  notes?: string;
  platforms?: GenerateDescriptionInputPlatformsItem[];
}