# lib/api-zod/src/generated/types/listingInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CrossList Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { ListingInputCondition } from './listingInputCondition';

export interface ListingInput {
  title: string;
  description?: string;
  price: number;
  originalPrice?: number;
  category?: string;
  brand?: string;
  size?: string;
  condition: ListingInputCondition;
  platforms?: string[];
  imageUrls?: string[];
}