# lib/api-zod/src/generated/types/platform.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CrossList Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface Platform {
  name: string;
  displayName: string;
  connected: boolean;
  color: string;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  listingCount?: number | null;
}