# lib/api-zod/src/generated/types/publishRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CrossList Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { PublishRequestPlatformsItem } from './publishRequestPlatformsItem';

export interface PublishRequest {
  platforms: PublishRequestPlatformsItem[];
}