# lib/api-zod/src/generated/types/publishResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CrossList Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { PlatformPublishResult } from './platformPublishResult';

export interface PublishResult {
  results: PlatformPublishResult[];
}