# lib/api-zod/src/generated/types/removeBgInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CrossList Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface RemoveBgInput {
  imageBase64: string;
  filename?: string;
}