# lib/api-zod/src/generated/types/removeBgResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CrossList Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface RemoveBgResult {
  imageBase64: string;
  filename: string;
}